Convert Crypto::Scalar to module, add .mult
This commit reimplements Crypto::Scalar as a module as that seems like a
more sensible API per the previous discussion.

It also adds a Crypto::Scalar.mult function that takes a group element
and an integer for doing things like Diffie-Hellman.

In addition, it pulls Crypto::Scalar.mult_base into pure Ruby, adding a
constant for the standard group element used to compute NaCl's
Curve25519 public keys. This means we only need to use an FFI binding
for .mult