Crypto::Serializable
This provides a module-based (as opposed to inheritance-based) approach
to serializing key-like things, including Crypto::Point.

Inheritance is cool and all, but I'm not sure it's the best choice here.
I kind of like modules in this application. Not sure "Serializable" is
the best name but feel free to bikeshed all over that ;)