Set encoding: binary for all files.
This fixes the encoding related errors on ruby 2.0 but I'm not sure
whether its a 'true fix' or just a hack that fixes the specs.
Regardless, I think makes sense for a crypto library to work with binary
encoding as much as possible.  In line with this, the box/open methods
are changed slightly such that they will return BINARY encoded buffers.