Have SecretBox thunk directly to XPoly1305
This is a strawman of how I think this should probably work. It
eliminates all the superfluous thunking that was going on between
SecretBox and SecretBox::XPoly1305.

At the very least, I feel that code would be better implemented with
Forwardable and def_delegator if this approach isn't taken.

Looking at this, I almost feel like SecretBox could service as a
"friendly" wrapper which handles things like encodings in a single
place, which would address earlier concerns that handling encoding isn't
DRY and that the primitives shouldn't provide those features.