Remove the encoding system
After much discussion it was agreed that the encoding system was out of place in
this library.

This commit removes it, and hopefully improves the general quality of the
codebase as well. All encoding-related thunks and glue code have been removed,
the way TestVectors work have been standardized, and any stray test vectors
still lingering inside of the tests have been moved to test_vectors.rb

Fixes #54