Add fallback support to `sodium_constant` function.
If `sodium_constant` attempts to retrieve a constant that the linked
version of libsodium does not define, FFI will raise an
`FFI:NotFoundError` when `attach_function` is called.

This commit changes the signature of `sodium_constant` to utilize
keyword arguments for the `name` argument and a newly added 'fallback'
argument. If `fallback` is non-nil, `sodium_primitive` will rescue the
`FFI::NotFoundError` and instead substitute a function that returns
the value provided by `fallback`.

This allows us to default to a reasonable value while still keeping
up-to-date if newer versions of libsodium change the constant.

Provides the functionality to address issue #183 cleanly.